+++ packages/cli/core/src/index.ts
@@ -5,12 +5,13 @@ import {
   ensureAbsolutePath,
   logger,
   INTERNAL_PLUGINS,
+  DEFAULT_SERVER_CONFIG,
 } from '@modern-js/utils';
 import { enable } from '@modern-js/plugin/node';
 import type { Hooks } from '@modern-js/types';
 import { ErrorObject } from 'ajv';
 import { initCommandsMap } from './utils/commander';
-import { resolveConfig, loadUserConfig } from './config';
+import { resolveConfig, loadUserConfig, addServerConfigToDeps } from './config';
 import { loadPlugins } from './loadPlugins';
 import {
   AppContext,
@@ -73,6 +74,7 @@ const initAppDir = async (cwd?: string): Promise<string> => {
 
 export interface CoreOptions {
   configFile?: string;
+  serverConfigFile?: string;
   packageJsonConfig?: string;
   plugins?: typeof INTERNAL_PLUGINS;
   onSchemaError?: (error: ErrorObject) => void;
@@ -83,6 +85,17 @@ export interface CoreOptions {
     sharedDir?: string;
   };
 }
+
+export const mergeOptions = (options?: CoreOptions) => {
+  const defaultOptions = {
+    serverConfigFile: DEFAULT_SERVER_CONFIG,
+  };
+
+  return {
+    ...defaultOptions,
+    ...options,
+  };
+};
 
 const createCli = () => {
   let hooksRunner: HooksRunner;
@@ -95,32 +108,42 @@ const createCli = () => {
 
     manager.clear();
 
-    restartOptions = options;
+    const mergedOptions = mergeOptions(options);
+
+    restartOptions = mergedOptions;
 
     const appDirectory = await initAppDir();
 
     initCommandsMap();
 
-    const metaName = options?.options?.metaName ?? 'MODERN';
+    const metaName = mergedOptions?.options?.metaName ?? 'MODERN';
     loadEnv(appDirectory, process.env[`${metaName.toUpperCase()}_ENV`]);
 
     const loaded = await loadUserConfig(
       appDirectory,
-      options?.configFile,
-      options?.packageJsonConfig,
+      mergedOptions?.configFile,
+      mergedOptions?.packageJsonConfig,
     );
 
     const plugins = loadPlugins(appDirectory, loaded.config, {
-      internalPlugins: options?.plugins,
+      internalPlugins: mergedOptions?.plugins,
     });
 
     plugins.forEach(plugin => plugin.cli && manager.usePlugin(plugin.cli));
 
-    const appContext = initAppContext(
+    const appContext = initAppContext({
       appDirectory,
       plugins,
-      loaded.filePath,
-      options?.options,
+      configFile: loaded.filePath,
+      options: mergedOptions?.options,
+      serverConfigFile: mergedOptions?.serverConfigFile,
+    });
+
+    // 将 server.config 加入到 loaded.dependencies，以便对文件监听做热更新
+    addServerConfigToDeps(
+      loaded.dependencies,
+      appDirectory,
+      mergedOptions.serverConfigFile,
     );
 
     manager.run(() => {
@@ -226,3 +249,9 @@ const createCli = () => {
 export const cli = createCli();
 
 export { initAppDir, initAppContext };
+
+declare module '@modern-js/utils/compiled/commander' {
+  export interface Command {
+    commandsMap: Map<string, Command>;
+  }
+}