+++ packages/cli/plugin-garfish/src/cli/index.ts
@@ -75,6 +75,11 @@ export default ({
             },
           },
           tools: {
+            devServer: {
+              headers: {
+                'Access-Control-Allow-Origin': '*',
+              },
+            },
             webpack: (
               webpackConfig: any,
               {
@@ -85,9 +90,6 @@ export default ({
             ) => {
               if (resolveOptions?.deploy?.microFrontend) {
                 chain.output.libraryTarget('umd');
-                chain.devServer.headers({
-                  'Access-Control-Allow-Origin': '*',
-                });
                 if (resolveOptions?.server?.port) {
                   chain.output.publicPath(
                     env === 'development'