+++ packages/server/core/tests/server-plugin.test.ts
@@ -1,4 +1,5 @@
 import { serverManager } from '../src';
+import { ServerConfig } from '../src/plugin';
 
 describe('Default cases', () => {
   it('Have returns plugins', async () => {
@@ -18,5 +19,42 @@ describe('Default cases', () => {
     await runner.prepareApiServer({ pwd: '', mode: 'function', config: {} });
 
     expect(count).toBe(1);
+  });
+
+  it('config hook should works correctly', async () => {
+    const proxy = {
+      '/bff': {
+        target: 'https://modernjs.dev',
+        changeOrigin: true,
+      },
+    };
+    const expectedServerConfig = {
+      bff: {
+        proxy,
+      },
+    };
+    let receivedServerConfig: ServerConfig = {};
+
+    serverManager.usePlugin(
+      serverManager.createPlugin(() => ({
+        config(serverConfig) {
+          serverConfig.bff = {
+            proxy,
+          };
+          return serverConfig;
+        },
+      })),
+
+      serverManager.createPlugin(() => ({
+        config(serverConfig) {
+          receivedServerConfig = serverConfig;
+          return receivedServerConfig;
+        },
+      })),
+    );
+
+    const runner = await serverManager.init();
+    runner.config({});
+    expect(expectedServerConfig).toEqual(receivedServerConfig);
   });
 });