+++ packages/toolkit/utils/src/index.ts
@@ -29,3 +29,5 @@ export * from './watch';
 export * from './nodeEnv';
 export * from './wait';
 export * from './emptyDir';
+export * from './getServerConfig';
+export * from './ssrHelpers';