+++ tests/integration/testing-bff/modern.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from '@modern-js/app-tools';
-
-export default defineConfig({
-  runtime: {
-    router: true,
-    state: true,
-  },
-});